Pokemon backend API tested
This PR includes several improvements to the project, such as updating dependencies, renaming the application class (removing the final suffix), and refactoring controller tests for better clarity and accuracy. The model file was cleaned up by removing unnecessary comments, and a README was added for documentation. The package name was fixed, and the project setup was adjusted to ensure smooth operation. Test assertions were enhanced to verify the correctness of returned data, and the service tests were reorganized for better clarity. Additionally, List<String> was replaced with List<PokemonType> to prevent future issues, the id property was removed as it was redundant, and the model directory name was updated to follow naming conventions. The Mockk library was also removed, and the test for retrieving all Pokémon was fixed. The full list of 151 Pokémon was added, represented by an enum class, and unnecessary imports and comments were cleaned up.